padding 계산 int 표현 지워보기

+++ src/userprog/process.c
@@ -111,10 +111,10 @@ void argument_passing(int argc, char **argv, struct intr_frame *_if){
     memcpy(_if->esp, argv[i], (int)strlen(argv[i]) + 1);
     argv[i] = (char*)_if->esp;
   }
-
-  printf("padding: %d", (int)_if->esp % 4);
-  _if->esp -= ((int)_if->esp % 4 + 4); //padding + argv[4]에 0 push
-  memset(_if->esp, 0, (int)_if->esp % 4 + 4);
+  
+  printf("padding: %d", _if->esp % 4);
+  _if->esp -= (_if->esp % 4 + 4); //padding + argv[4]에 0 push
+  memset(_if->esp, 0, _if->esp % 4 + 4);
   printf("************padding + argv[4] push\nesp: %x\n**************\n", (int)_if->esp);
 
   for(int i = argc - 1; i >= 0; i--){ //argv[] address push
